relocate directory hierarchy
