+++ src/main/java/com/C_platform/Member/entity/UserRole.java
@@ -1,0 +1,16 @@
+package com.C_platform.Member.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum UserRole {
+    ADMIN("ROLE_ADMIN"),
+    USER("ROLE_USER");
+
+    private String value;
+
+    UserRole(String value) {
+        this.value = value;
+    }
+}
+
