+++ src/main/java/com/C_platform/Member/entity/loginType.java
@@ -1,4 +1,4 @@
-package com.C_platform.item.entity.Member;
+package com.C_platform.Member.entity;
 
 // CUSTOM은 spring login 기능을 이용하여 직접 login을 구현하는 방법
 public enum loginType {
