+++ src/main/java/com/C_platform/Member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package com.C_platform.Member.repository;
+
+import com.C_platform.Member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Integer> {
+}
