+++ src/main/java/com/C_platform/Member/service/MemberService.java
@@ -1,0 +1,11 @@
+package com.C_platform.Member.service;
+
+import com.C_platform.Member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberService {
+    private final MemberRepository memberRepository;
+}
