+++ src/main/java/com/C_platform/controller/MemberController.java
@@ -1,5 +1,5 @@
 package com.C_platform.controller;
 
-
-public class UserController {
+@
+public class MemberController {
 }
