+++ src/main/java/com/C_platform/order/Order.java
@@ -1,5 +1,6 @@
-package com.C_platform.item.entity.Member;
+package com.C_platform.order;
 
+import com.C_platform.Member.entity.Member;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
